automake: Disable portability warnings
When they are enabled, autogen.sh fails with loads of warnings on F19.

-Wportability used to be automatically disabled with automake <= 1.12
when using silent rules, but this is no longer the case with automake
1.13 which is what fedora 19 uses:

http://www.flameeyes.eu/autotools-mythbuster/automake/silent.html

"As of version 1.13, though, this opt-in is no longer necessary, as all
the Makefiles are generated to support them. The silent-rules option is
now a no-op, doing nothing at all, in particular not silencing the
portability warnings."

This commit disables these warnings in order to avoid autogen.sh
breakage because of the use of -Werror, they can be reenabled once
gtk-doc.make is fixed to avoid these portability warnings.

This patch and commit log is based on a patch from Christophe Fergeau
<cfergeau@redhat.com> to spice-gtk.