service-manager: Give agents time to register on autolaunch
If geoclue was auto-launched on a Manager.GetClient() call, agents
haven't yet got the time to register and we'll end up falsely
denying apps access to geolocation data.

So if there is no agent for appropriate user on a call to Manager.GetClient(),
we wait 1 second before replying to client app.