config: Allow whitelisting of applications
Some applications are part of the system and/or ensure that user
location is only queried with their consent. Examples are
gnome-settings-daemon and web browsers. We setup a small list of these
by default and allow system admins/packagers to edit this list through
our config file.

Later we can re-use the same format (or even the file) to add/remove
apps to whitelist based on agents' verdict.