wifi,config: Use an API key
To be able to counter the attacks on its service, Mozilla folks now
require apps to provide a key. At the moment, this could be anything but
in future, they'll allocate us a key and we'll have to use that. Lets
keep it in the configuration file and when/if we need to change it in
future, people can make this working again but changing this in their
config file rather than be forced to update to latest geoclue (assuming
their distro will even provide an update).