wifi,config: Easy way to submit WiFi data to Mozilla
Provide an option in config file to enable submission of wifi data to
Mozilla Location Service. If this option is enabled, the data is submitted
each time geoclue get a GPS lock. That means, its not completely
autonomous and data is only submitted if an application, e.g gnome-maps
requests current location with exact accuracy.