Add support for disabling some sources
Introduce configure options --disable-wifi-source, --disable-3g-source
and --disable-modem-gps-source to allow disabling of backends that
depend on NetworkManager and ModemManager (since these components are
not widely ported).

There is no "auto detect" -- the options must be explicitly specified in
order to disable the feature.  This prevents a quiet reduction in
functionality just because the right headers didn't happen to be
installed at configure time, which should help prevent packaging
mistakes.

https://bugs.freedesktop.org/show_bug.cgi?id=73961