Drop now redundant geoip server
The plan was to put this server running on a GNOME server but Mozilla
Location Service starting to support geoip as well, made us change that
plan. Now that we are using that and have dropped our standalone
ipclient, we can't even use this server readily any ways so there is no
point in keeping it around.