modem-source: Don't disable modem on stopping
Disabling the modem, disables it for everyone so if there was an active
connection going on through the modem, geoclue will cause it to
disconnect. While we need to have some centralized means to
enable/disable modem and ideally geoclue shouldn't do any
enabling/disabling of modems at all, for now lets just only enable the
modem.

Also turns out that modem disabling itself doesn't affect the power
consumption and we'll need an extra call to put modem to low power
usage.