service-manager: 'InUse' unaffected by system components
Lets keep this property for apps only and therefore don't report
location in use for system components. With this change, gnome-shell
won't permanently show location icon in tray if automatic timezone
is enabled.

One thing this patch breaks is not to timeout if service is in use by
system components. It'll be fixed in the following patches.