locator: Most accurate source first
We now maintain the list of sources sorted based on their accuracy
level. This allows to start the source providing the most accurate
location first. When multiple sources are already active due to the
activity of an existing client, the start of another client will
hopefully give it the best possible location only. It also works better
when clients work in "single shot" mode, and use only the first location
received from geoclue.

https://bugs.freedesktop.org/show_bug.cgi?id=89719