location-source: Always create new location instance
The timestamp on location is only set on it during creation so if we
re-use the same location object, we'll have the same timestamp on every
new location and hence we won't be able to calculate speed by comparing
the new and old locations.

Besides the timestamp is supposed to represent the time when location
was acquired so keeping it the same throughout the lifetime of geoclue,
is just wrong.

https://bugs.freedesktop.org/show_bug.cgi?id=89852