locator: Move speed/heading setting to source
Move calculation and setting of speed and heading to
gclue_location_source_set_location(). We ensure that speed and heading
are only calculated if they are not already set on the new location
object so that they are not re-calculated each time location is set from
one source to another (keeping in mind that Locator is also a
LocationSource subclass).

This change would mean that speed and heading are only set by the source
it originates from. i-e speed and heading are not calculated by comparing
location from two different sources, which would typically be wrong as
different locations from different sources just means the same location
but with different accuracies. For example, WiFi source will put me at
Crystal Palace but soon after GPS will put me at Westow House (a pub in
Crystal Palace) but I'm not moving so deriving speed and heading from
these two locations would be wrong.

As a positive side-effect, this would make it possible for sources to
provide the speed and accuracy themselves. This would be at least
possible for GPS (see bug#89907).

Known issue: We still end up with these properties being computed by
Locator by comparison of locations from two different sources if the
latest location is the first location from a source. We solve this issue
in a following patch.

https://bugs.freedesktop.org/show_bug.cgi?id=89852