locator: Don't auto-compute speed & heading
Now that the actual sources themselves ensure that speed and heading are
computed and set on new locations whenever possible, there is no reason
for Locator to do this at all anymore. Otherwise, we still end up with
these properties being computed by comparison of locations from two
different sources if the latest location is the first location from a
source.

https://bugs.freedesktop.org/show_bug.cgi?id=89852